docs(README): added Table of Contents (ToC) for the Paraduxical project that links to all of the individual headers/subsections of our README.md (i.e., the main document for the project)
I still need to test the ToC Markdown on GitLab to make sure their flavor of Markdown works with what we have for documentation. I had to push it to this branch and commit the changes because there is no way that I am aware of to test GitLab's flavor of Markdown locally or in my local development environment. I will make futher commits to address any problems that may come out of GitLab's Markdown processor and renderer not being compatible with our current README.md to make sure everything works.