feat: user-defined ports now correctly propogate to the back-end, TUI, and GUI stacks
Whenever the user defines any port from the CLI, it will now correctly propagate to the TUI and GUI stacks of the project. The back-end will also use the user-defined port as its listening/hosting port, so there are no worries there.

Bonus: the issue with the GUI throwing exceptions about the URL it is supposed to connect to being incorrect got fixed with this commit, which is due to how the user-defined ports get sent down to the correct classes from the controller and the main/runner class.